Add update steps to manage interviews scenario
